Hyungwoo branch

+++ src/main/java/com/treemarket/tree/service/CategoryService.java
@@ -1,5 +1,5 @@
 package com.treemarket.tree.service;
 
 public interface CategoryService {
-    int getCtgId (String ctgName);
+    Long getCtgId (String ctgName);
 }
